+++ tests/compile-fail/wildcard_before_input_state.rs
@@ -1,0 +1,12 @@
+extern crate smlang;
+
+use smlang::statemachine;
+
+statemachine! {
+    transitions: {
+        _ + Event1 = Fault, //~ State and event combination specified multiple times, remove duplicates.
+        *State1 + Event1 = State2,
+    }
+}
+
+fn main() {}