+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 
 - Starting state can now contain data ([issue-34](https://github.com/korken89/smlang-rs/issues/34))
+- Allow explicit input states before wildcard input state([issue-47](https://github.com/korken89/smlang-rs/pull/47)
 
 ## [v0.5.1]